Ensure imports only from lazy-syntaxhighlighter
