+++ code/core/src/components/components/syntaxhighlighter/syntaxhighlighter.tsx
@@ -26,13 +26,14 @@ import { styled } from 'storybook/theming';
 import { ActionBar } from '../ActionBar/ActionBar';
 import type { ScrollAreaProps } from '../ScrollArea/ScrollArea';
 import { ScrollArea } from '../ScrollArea/ScrollArea';
+import { createCopyToClipboardFunction } from './clipboard';
 import type {
   SyntaxHighlighterProps,
   SyntaxHighlighterRenderer,
   SyntaxHighlighterRendererProps,
 } from './syntaxhighlighter-types';
 
-const { document, window: globalWindow } = global;
+const { window: globalWindow } = global;
 
 export const supportedLanguages = {
   jsextra: jsExtras,
@@ -56,31 +57,7 @@ const themedSyntax = memoize(2)((theme) =>
   Object.entries(theme.code || {}).reduce((acc, [key, val]) => ({ ...acc, [`* .${key}`]: val }), {})
 );
 
-const copyToClipboard: (text: string) => Promise<void> = createCopyToClipboardFunction();
-
-export function createCopyToClipboardFunction() {
-  if (globalWindow.navigator?.clipboard) {
-    return async (text: string) => {
-      try {
-        await globalWindow.top?.navigator.clipboard.writeText(text);
-      } catch {
-        await globalWindow.navigator.clipboard.writeText(text);
-      }
-    };
-  }
-  return async (text: string) => {
-    const tmp = document.createElement('TEXTAREA') as HTMLTextAreaElement;
-    const focus = document.activeElement as HTMLTextAreaElement;
-
-    tmp.value = text;
-
-    document.body.appendChild(tmp);
-    tmp.select();
-    document.execCommand('copy');
-    document.body.removeChild(tmp);
-    focus.focus();
-  };
-}
+const copyToClipboard: (text: string) => Promise<void> = createCopToClipboardFunction();
 
 export interface WrapperProps {
   bordered?: boolean;
